WIP: Auth with API Key
In terms of the actual application code, things are pretty straightforward: if an API key is passed, find the
associated user, check the hash against the passed API key, and attach the user if everything went well.

The hackiest bit of this is in the test code. Our HTTP client function takes a lot of positional arguments, many of
which are optional, so adding in a header to an arbitrary call is not easy. Luckily, it's easy to parse the args,
change what we want, and then unparse them before actually calling the client - but I'm not sure how I feel about
this. Part of me thinks it might be better to have a new HTTP client function that takes a map instead of all these
positional args...

I do want to rename the existing functions related to API keys, e.g. `mw.auth/wrap-api-key`. We should be careful
about overloading the term "API Key", both internally and externally.

The External "API Key" is a static key defined by an env var that allows you to require a particular header *in
addition to regular authentication* (on some endpoints).

It's probably too painful to change the actual setting name, but we can change what we call it in the code.

I also need to add more/better tests.