Add a name to API keys
Instead of reusing the `core_user.last_name` column for the API key
name, it makes more sense to have an actual column that represents it.
For one thing, it allows us to have an actual UNIQUE index on the field
and assert that it's non-nullable at the database layer. It also just
seems better to not overload a column with multiple semantic types of
data.